requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiohttp-swagger==1.0.15
async-lru==1.0.2
async-timeout==3.0.1
asyncssh==2.7.2
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
cryptography==35.0.0
dill==0.3.4
distro==1.6.0
genie==21.9
genie.libs.clean==21.9.3
genie.libs.conf==21.9
genie.libs.filetransferutils==21.9
genie.libs.health==21.9.1
genie.libs.ops==21.9
genie.libs.parser==21.9
genie.libs.sdk==21.9
idna==3.3
Jinja2==2.11.3
jsonpickle==2.0.0
junit-xml==1.9
MarkupSafe==1.1.1
more-itertools==8.10.0
multidict==5.2.0
netaddr==0.8.0
packaging==21.0
pathspec==0.9.0
pip==21.2.4
pluggy==0.13.1
prettytable==2.2.1
psutil==5.8.0
py==1.10.0
pyats==21.9
pyats.aereport==21.9
pyats.aetest==21.9
pyats.async==21.9
pyats.connections==21.9
pyats.datastructures==21.9
pyats.easypy==21.9
pyats.kleenex==21.9
pyats.log==21.9
pyats.reporter==21.9
pyats.results==21.9
pyats.tcl==21.9
pyats.topology==21.9
pyats.utils==21.9
pycparser==2.20
pyftpdlib==1.5.6
pyparsing==2.4.7
pytest==5.4.3
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scrapli==2021.7.30
scrapli-community==2021.7.30
setuptools==58.1.0
six==1.16.0
ssh2-python==0.26.0
tftpy==0.8.0
tqdm==4.62.3
typing-extensions==3.10.0.2
unicon==21.9
unicon.plugins==21.9
urllib3==1.26.7
wcwidth==0.2.5
wheel==0.37.0
xmltodict==0.12.0
yamllint==1.26.3
yarl==1.7.0
